Split into two services and add a timer
Currently fake-hwclock is packaged as a single systemd service which
restores the clock time on startup and saves the clock time on shutdown,
using "ExecStart=" and "ExecStop=" configuration options. There's also a
cron script to run the save operation periodically in case the system
shuts down ungracefully, but it would be better to provide a systemd
timer so that systems which don't have cron still have that
functionality. However such a timer couldn't activate the fake-hwclock
service as written because (a) it has RemainAfterExit=yes and is
therefore always active and (b) the timer would be restoring and
saving the time periodically, rather than just saving it as intended.

So this commit breaks up fake-hwclock.service into two services:
fake-hwclock-restore.service and fake-hwclock.service, latter of which
does save operations. The restore service is set up to run on startup
and the save service to run on shutdown, so that the current behavior is
maintained.

This commit also adds a systemd timer unit with the same purpose as the
cron script, to save the clock time to disk periodically.

The save service and timer are called fake-hwclock.{service,timer} rather than
fake-hwclock-save.{service,timer} so they shadow the LSB init script of
the same name, preventing systemd-sysv-generator from generating an
(inert) fake-hwclock.service. This also means they don't have to be
explicitly listed in .install, since $package.service and $package.timer
are automatically included in $package if present.